intiation

+++ database/migrations/2020_08_25_191944_create_services_table.php
@@ -16,7 +16,7 @@ class CreateServicesTable extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->id();
             $table->string("name");
-            $table->float("price");
+            $table->integer("price");
             $table->timestamps();
         });
     }
